+++ CHANGELOG.md
@@ -4,7 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.0] - 2020-06-25
+## [1.1.0] - 2020-08-22
+### Added
+- Custom ValidatorException
+- LICENSE file
+### Changed
+- Main plugins now throw ValidatorException
+
+## [1.0.0] - 2020-08-18
 ### Added
 - Compatibility chart
 ### Changed